python-app/requirements.txt
fastapi
uvicorn[standard]
python-jose
passlib[bcrypt]
jinja2
prometheus_client
python-dotenv
anyio
